+++ nikonov.andrew/T1/diamond.hpp
@@ -8,11 +8,11 @@ namespace nikonov
   {
   public:
     Diamond(point_t topP, point_t bottomP, point_t leftP);
-    float getArea() const override;
+    double  getArea() const override;
     rectangle_t getFrameRect() const override;
     void move(point_t a) override;
-    void move(float x, float y) override;
-    void scale(float k) override;
+    void move(double  x, double  y) override;
+    void scale(double  k) override;
   private:
     point_t topP_, rightP_, midP_;
   };
