+++ nikonov.andrew/T1/shape.hpp
@@ -5,11 +5,11 @@ namespace nikonov
 {
   struct Shape
   {
-    virtual float getArea() const = 0;
+    virtual double  getArea() const = 0;
     virtual nikonov::rectangle_t getFrameRect() const = 0;
     virtual void move(point_t a) = 0;
-    virtual void move(float x, float y) = 0;
-    virtual void scale(float k) = 0;
+    virtual void move(double  x, double  y) = 0;
+    virtual void scale(double  k) = 0;
     virtual ~Shape() = default;
   };
 }
